+++ source/Animation.cpp
@@ -20,12 +20,13 @@ Animation::Animation()
  * @param count number of frames in texture
  * @param speed animation speed
  */
-Animation::Animation(const sf::Texture& t, const Vector<float>& position,
+Animation::Animation(const sf::Texture& t, const Vector<int>& position,
                      const int& w, const int& h, const int& count, const float& speed)
     : frame(0), speed(speed)
 {
     for(int i = 0; i < count; ++i)
-        frames.push_back(sf::IntRect(static_cast<int>(position.x + i * w), static_cast<int>(position.y), w, h));
+        frames.push_back(sf::IntRect(position.x + i * w, position.y, w, h));
+
 
     sprite.setTexture(t);
     sprite.setOrigin(static_cast<float>(w/2), static_cast<float>(h/2));
